requirements.txt
langchain
openai
python-dotenv
pydantic
llama-index